backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
django-filter==23.3
django-colorfield==0.9.0
django-chartjs==2.3.0
django-seed==0.3.1
drf-yasg==1.21.7
whitenoise==6.5.0
pytest==7.4.3
pytest-django==4.5.2
Pillow==10.1.0
python-dateutil==2.8.2
